feat(logistics/start): se integró los valores de App Params en la functions...

See merge request apps/operaciones-industriales/blending-harina-2/blending-harina-2-backend!22

+++ Function.Blending.Opt/Infrastructure/Persistence/Models/LogInpFiltro.cs
@@ -19,7 +19,7 @@ public partial class LogInpFiltro
 
     public string? Division { get; set; }
 
-    public int? TiempoEspera { get; set; }
+    public decimal? TiempoEspera { get; set; }
 
     public virtual LogEjecucion Ejecucion { get; set; } = null!;
